+++ docs/source/conf.py
@@ -597,6 +597,8 @@ def create_doctest_figure(app, obj, name, lines):
         sys.argv.append('--show')
     if '--nointeract' not in sys.argv:
         sys.argv.append('--nointeract')
+    if '--network' not in sys.argv:
+        sys.argv.append('--network')
     modpath = module.__file__
 
     # print(doctest.format_src())
@@ -654,7 +656,9 @@ def create_doctest_figure(app, obj, name, lines):
         num_lines = part.count('\n')
 
         doctests = list(xdoctest.core.parse_docstr_examples(
-            part, modpath=modpath, callname=name))
+            part, modpath=modpath, callname=name,
+            # style='google'
+        ))
         # print(doctests)
 
         # doctests = list(xdoctest.core.parse_docstr_examples(